+++ templates/auv3-cmake/lib/domain/interfaces.h
@@ -20,21 +20,21 @@ public:
 class IParametersStore {
 public:
   virtual ~IParametersStore() = default;
-  virtual double get(uint32_t id) = 0;
-  virtual void set(uint32_t id, double value) = 0;
+  virtual float get(uint32_t id) = 0;
+  virtual void set(uint32_t id, float value) = 0;
 };
 
 class IParameterManager {
 public:
   virtual ~IParameterManager() = default;
 
-  virtual void setParameter(ParamId id, double value, bool notifyToUi) = 0;
-  virtual double getParameter(ParamId id) = 0;
+  virtual void setParameter(ParamId id, float value, bool notifyToUi) = 0;
+  virtual float getParameter(ParamId id) = 0;
 
-  virtual void getAllParameters(std::map<std::string, double> &parameters) = 0;
+  virtual void getAllParameters(std::map<std::string, float> &parameters) = 0;
 
   virtual int subscribeParameterChange(
-      std::function<void(const std::string paramKey, double value)>
+      std::function<void(const std::string paramKey, float value)>
           callback) = 0;
   virtual void unsubscribeParameterChange(int subscriptionId) = 0;
 };
@@ -49,21 +49,21 @@ enum class ParameterEditState {
 // class IParameterEditActionDestinationPort {
 // public:
 //   virtual ~IParameterEditActionDestinationPort() = default;
-//   virtual void applyParameterEditFromUi(std::string paramKey, double value,
+//   virtual void applyParameterEditFromUi(std::string paramKey, float value,
 //                                         ParameterEditState editState) = 0;
 // };
 
 class IControllerFacade {
 public:
   virtual ~IControllerFacade() = default;
-  virtual void getAllParameters(std::map<std::string, double> &parameters) = 0;
-  virtual void applyParameterEditFromUi(std::string paramKey, double value,
+  virtual void getAllParameters(std::map<std::string, float> &parameters) = 0;
+  virtual void applyParameterEditFromUi(std::string paramKey, float value,
                                         ParameterEditState editState) = 0;
-  virtual void requestNoteOn(int noteNumber, double velocity) = 0;
+  virtual void requestNoteOn(int noteNumber, float velocity) = 0;
   virtual void requestNoteOff(int noteNumber) = 0;
 
   virtual int subscribeParameterChange(
-      std::function<void(const std::string paramKey, double value)>
+      std::function<void(const std::string paramKey, float value)>
           callback) = 0;
   virtual void unsubscribeParameterChange(int subscriptionId) = 0;
 };
